refactor(channels): shared preview kit — one failure shape, one wording, one arg summary
TelegramFailure/FeishuFailure, defaultErrorMessage and summarizeArgs
were verbatim (or near-verbatim) copies across the two previews — the
customer-facing wording could drift channel by channel. Extracted to
channels/preview-kit.ts (ChannelFailure + defaultErrorMessage +
summarizeToolArgs); the platform-branded Failure types stay as public
aliases (the LarkFailure precedent). The kit's clip takes Feishu's
code-point-safe version (Telegram's naive slice could tear a surrogate
pair mid-emoji), so feishu/text.ts moves up to channels/text.ts — it
was platform-neutral all along. Preview lifecycles stay per-platform.
