+++ AGENTS.md
@@ -45,6 +45,8 @@ src/
 │   ├── http.ts              # HTTP/SSE channel (consumes only the Agent contract)
 │   ├── control.ts           # session-control transport: bearer-token /control/* routes (dispatch + SSE events with wire envelope + /control/invoke)
 │   ├── body.ts, respond.ts  # channel-authoring kit (body cap, responses)
+│   ├── preview-kit.ts       # SHARED preview policies: ChannelFailure + customer-facing default error + tool-arg summary
+│   ├── text.ts              # SHARED Unicode-safe code-point slicing (cards, preview kit)
 │   ├── turn-queue.ts        # SHARED: in-memory per-session serial turns (FIFO; telegram + feishu)
 │   ├── turn-store.ts        # SHARED: generic durable turn intent (L1) — record shape/validator/order injected per channel
 │   ├── state.ts             # SHARED: atomic state files under <stateRoot>/channels/<kind>/
@@ -68,7 +70,6 @@ src/
 │   │   ├── invoke-turn.ts, preview.ts, seen.ts # turn IO, streaming-card delivery, delivery dedup
 │   │   ├── owned-threads.ts # durable index of Agent-created group threads (admits bare continuations)
 │   │   ├── context-buffer.ts# unsummoned group/thread discussion (durable, commit-on-completed)
-│   │   ├── text.ts          # Unicode-safe code-point slicing for card/text rendering
 │   │   ├── feishu-api.ts    # canonical Open API pipeline (token cache, retry, cardkit)
 │   │   ├── register-app.ts  # `add feishu`: scan-to-create device flow
 │   │   ├── register-webhook.ts, bootstrap-token.ts # event URL + token automation
