+++ src/channels/feishu/card.ts
@@ -12,7 +12,7 @@
  * well under it; longer answers overflow into follow-up messages (preview.ts owns that policy).
  */
 
-import { truncateCodePointPrefix } from "./text.ts";
+import { truncateCodePointPrefix } from "../text.ts";
 
 /** The one streamed element's id — shared by create (card.ts) and update (preview.ts). */
 export const ANSWER_ELEMENT_ID = "answer";
